+++ src/main/java/org/plumelib/util/EntryReader.java
@@ -9,6 +9,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.LineNumberReader;
 import java.io.Reader;
+import java.io.UncheckedIOException;
 import java.io.UnsupportedEncodingException;
 import java.nio.CharBuffer;
 import java.nio.file.Path;
@@ -21,6 +22,7 @@ import org.checkerframework.checker.index.qual.GTENegativeOne;
 import org.checkerframework.checker.index.qual.IndexOrLow;
 import org.checkerframework.checker.index.qual.NonNegative;
 import org.checkerframework.checker.lock.qual.GuardSatisfied;
+import org.checkerframework.checker.mustcall.qual.MustCallAlias;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -68,7 +70,10 @@ import org.checkerframework.checker.regex.qual.Regex;
  *
  * @see #getEntry() and @see #setEntryStartStop(String,String)
  */
-@SuppressWarnings({"IterableAndIterator"})
+@SuppressWarnings({
+  "IterableAndIterator",
+  "builder:required.method.not.called" // Collection `readers` has element type @MustCall("close")
+})
 public class EntryReader extends LineNumberReader implements Iterable<String>, Iterator<String> {
 
   ///
@@ -221,8 +226,8 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *     should define one group that contains the include file name.
    * @see #EntryReader(InputStream,String,String,String)
    */
-  public EntryReader(
-      InputStream in,
+  public @MustCallAlias EntryReader(
+      @MustCallAlias InputStream in,
       String charsetName,
       String filename,
       @Nullable @Regex String commentRegexString,
@@ -240,7 +245,8 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @throws UnsupportedEncodingException if the charset encoding is not supported
    * @see #EntryReader(InputStream,String,String,String)
    */
-  public EntryReader(InputStream in, String charsetName, String filename)
+  public @MustCallAlias EntryReader(
+      @MustCallAlias InputStream in, String charsetName, String filename)
       throws UnsupportedEncodingException {
     this(in, charsetName, filename, null, null);
   }
@@ -257,8 +263,8 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param includeRegexString regular expression that matches include directives. The expression
    *     should define one group that contains the include file name.
    */
-  public EntryReader(
-      InputStream in,
+  public @MustCallAlias EntryReader(
+      @MustCallAlias InputStream in,
       String filename,
       @Nullable @Regex String commentRegexString,
       @Nullable @Regex(1) String includeRegexString) {
@@ -273,7 +279,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param filename the file name
    * @see #EntryReader(InputStream,String,String,String,String)
    */
-  public EntryReader(InputStream in, String filename) {
+  public @MustCallAlias EntryReader(@MustCallAlias InputStream in, String filename) {
     this(in, filename, null, null);
   }
 
@@ -283,12 +289,24 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param in the InputStream
    * @see #EntryReader(InputStream,String,String,String)
    */
-  public EntryReader(InputStream in) {
+  public @MustCallAlias EntryReader(@MustCallAlias InputStream in) {
     this(in, "(InputStream)", null, null);
   }
 
   /** A dummy Reader to be used when null is not acceptable. */
   private static class DummyReader extends Reader {
+
+    /** The canonical DummyReader. */
+    public static final DummyReader it = new DummyReader();
+
+    /**
+     * Create a new DummyReader.
+     *
+     * @deprecated use {@link #it}.
+     */
+    @Deprecated // 2022-07-25; to make private
+    public DummyReader() {}
+
     @Override
     public void close(@GuardSatisfied DummyReader this) {
       // No error, because closing is OK if it appears in try-with-resources.
@@ -352,14 +370,15 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param includeRegexString regular expression that matches include directives. The expression
    *     should define one group that contains the include file name
    */
-  public EntryReader(
-      Reader reader,
+  @SuppressWarnings("builder") // storing into a collection
+  public @MustCallAlias EntryReader(
+      @MustCallAlias Reader reader,
       String filename,
       @Nullable @Regex String commentRegexString,
       @Nullable @Regex(1) String includeRegexString) {
     // we won't use superclass methods, but passing null as an argument
     // leads to a NullPointerException.
-    super(new DummyReader());
+    super(DummyReader.it);
     readers.addFirst(new FlnReader(reader, filename));
     if (commentRegexString == null) {
       commentRegex = null;
@@ -379,7 +398,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    * @param reader source from which to read entries
    * @see #EntryReader(Reader,String,String,String)
    */
-  public EntryReader(Reader reader) {
+  public @MustCallAlias EntryReader(@MustCallAlias Reader reader) {
     this(reader, reader.toString(), null, null);
   }
 
@@ -573,7 +592,8 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
           // System.out.printf ("absolute filename = %s %s %s%n",
           //                     currentFilename, currentParent, filename);
         }
-        readers.addFirst(new FlnReader(filename.getAbsolutePath()));
+        FlnReader reader = new FlnReader(filename.getAbsolutePath());
+        readers.addFirst(reader);
         return readLine();
       }
     }
@@ -591,8 +611,9 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
    *
    * @return a line-by-line iterator for this file
    */
+  @SuppressWarnings("mustcall:override.return")
   @Override
-  public Iterator<String> iterator(EntryReader this) {
+  public @MustCallAlias Iterator<String> iterator(@MustCallAlias EntryReader this) {
     return this;
   }
 
@@ -616,7 +637,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
     try {
       line = readLine();
     } catch (IOException e) {
-      throw new Error("unexpected IOException: ", e);
+      throw new UncheckedIOException(e);
     }
 
     if (line == null) {
@@ -643,7 +664,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
         throw new NoSuchElementException();
       }
     } catch (IOException e) {
-      throw new Error("unexpected IOException", e);
+      throw new UncheckedIOException(e);
     }
   }
 
@@ -761,7 +782,7 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
     FlnReader ri1 = readers.getFirst();
     String line = ri1.readLine();
     while (line == null) {
-      readers.removeFirst();
+      readers.removeFirst().close();
       if (readers.isEmpty()) {
         return null;
       }
@@ -920,12 +941,13 @@ public class EntryReader extends LineNumberReader implements Iterable<String>, I
         System.exit(1);
       }
     }
-    EntryReader reader = new EntryReader(filename, commentRegex, includeRegex);
+    try (EntryReader reader = new EntryReader(filename, commentRegex, includeRegex)) {
 
-    String line = reader.readLine();
-    while (line != null) {
-      System.out.printf("%s: %d: %s%n", reader.getFileName(), reader.getLineNumber(), line);
-      line = reader.readLine();
+      String line = reader.readLine();
+      while (line != null) {
+        System.out.printf("%s: %d: %s%n", reader.getFileName(), reader.getLineNumber(), line);
+        line = reader.readLine();
+      }
     }
   }
 }